removes mathlib2 tests and refactors to mathlib
